improve replies
