+++ agent/src/temp/reply-testing.ts
@@ -10,9 +10,10 @@ import {
   SYSTEM_PROMPT,
   TWITTER_REPLY_TEMPLATE,
 } from '../twitter/prompts';
-import { generateEmbedding, getTweet } from '../twitter/helpers';
+import { getTweet } from '../twitter/helpers';
 import { REJECTION_REASON } from '../const';
 import { openai } from '@ai-sdk/openai';
+import { getBillContext, getTweetContext } from '../twitter/execute';
 dotenv.config();
 
 const xAi = createXai({});
@@ -45,118 +46,6 @@ const terminal = readline.createInterface({
   input: process.stdin,
   output: process.stdout,
 });
-
-/**
- * given a tweet id, we try to follow the thread get more context about the tweet
- */
-async function getTweetContext({ id }: { id: string }) {
-  let tweets: Awaited<ReturnType<typeof getTweet>>[] = [];
-
-  let searchId: null | string = id;
-  do {
-    const tweet = await getTweet({ id: searchId });
-    tweets.push(tweet);
-    if (tweet.inReplyToId) {
-      searchId = tweet.inReplyToId;
-    }
-
-    if (tweet.inReplyToId === null) {
-      searchId = null;
-    }
-
-    // Limit to 5 tweets
-    if (tweets.length > 5) {
-      searchId = null;
-    }
-  } while (searchId);
-
-  return tweets
-    .reverse()
-    .map(tweet => tweet.text)
-    .join('\n---\n');
-}
-
-/**
- * Given some text try to narrow down the bill to focus on.
- *
- * If you get a bill title.
- * then we can filter the embeddings to that title
- * and then try to search for user question for embeddings
- *
- * If you do not get a bill title
- * we are out of luck and we just use the user question to search all the embeddings
- */
-async function getBillContext({
-  text,
-  question,
-}: {
-  text: string;
-  question: string;
-}) {
-  const billTitleResult = await generateText({
-    model: openai('gpt-4o'),
-    temperature: 0,
-    messages: [
-      {
-        role: 'system',
-        content: `You are an AI specialized in analyzing tweets related to U.S. Congressional bills. Given a tweet, extract the official title of the bill mentioned. If multiple bills are referenced, list all their titles. If no bill is mentioned, respond with 'NO_TITLE_FOUND.' Return only the title(s) without additional commentary.`,
-      },
-      {
-        role: 'user',
-        content: text,
-      },
-    ],
-  });
-
-  const questionEmbedding = await generateEmbedding(question);
-  const embeddingArrayString = JSON.stringify(questionEmbedding);
-
-  const billTitle = billTitleResult.text;
-
-  if (billTitle.startsWith('NO_TITLE_FOUND')) {
-    const vectorSearch = await db
-      .select({
-        text: billVector.text,
-        bill: billVector.bill,
-        distance: sql`vector_distance_cos(${billVector.vector}, vector32(${embeddingArrayString}))`,
-      })
-      .from(billVector)
-      .orderBy(
-        // ascending order
-        sql`vector_distance_cos(${billVector.vector}, vector32(${embeddingArrayString})) ASC`,
-      )
-      .limit(10);
-
-    return vectorSearch.map(row => row.text).join('\n---\n');
-  }
-
-  const billSearch = await db
-    .select({ id: bill.id })
-    .from(bill)
-    .where(like(bill.title, billTitle))
-    .limit(5);
-
-  const vectorSearch = await db
-    .select({
-      text: billVector.text,
-      bill: billVector.bill,
-      distance: sql`vector_distance_cos(${billVector.vector}, vector32(${embeddingArrayString}))`,
-    })
-    .from(billVector)
-    .where(
-      inArray(
-        billVector.bill,
-        billSearch.map(row => row.id),
-      ),
-    )
-    .orderBy(
-      // ascending order
-      sql`vector_distance_cos(${billVector.vector}, vector32(${embeddingArrayString})) ASC`,
-    )
-    .limit(10);
-
-  return vectorSearch.map(row => row.text).join('\n---\n');
-}
 
 async function main() {
   const tweetUrl = await terminal.question('Enter the tweet URL: ');
