+++ agent/src/twitter/execute.ts
@@ -6,16 +6,20 @@ import { createXai } from '@ai-sdk/xai';
 import { generateText, embedMany, embed } from 'ai';
 import { openai } from '@ai-sdk/openai';
 import {
+  EXTRACT_BILL_TITLE_PROMPT,
   QUESTION_EXTRACTOR_SYSTEM_PROMPT,
   SYSTEM_PROMPT,
   TWITTER_REPLY_TEMPLATE,
 } from './prompts';
 import { RecursiveCharacterTextSplitter } from '@langchain/textsplitters';
 import {
+  bill,
   billVector,
   chat as chatDbSchema,
   db,
   eq,
+  inArray,
+  like,
   message as messageDbSchema,
   messageVector,
   sql,
@@ -87,6 +91,122 @@ async function upsertChat({
 
   return chat[0];
 }
+
+/**
+ * given a tweet id, we try to follow the thread get more context about the tweet
+ */
+export async function getTweetContext({ id }: { id: string }) {
+  const LIMIT = 5;
+  let tweets: Awaited<ReturnType<typeof getTweet>>[] = [];
+
+  let searchId: null | string = id;
+  do {
+    const tweet = await getTweet({ id: searchId });
+    tweets.push(tweet);
+    if (tweet.inReplyToId) {
+      searchId = tweet.inReplyToId;
+    }
+
+    if (tweet.inReplyToId === null) {
+      searchId = null;
+    }
+
+    // Limit max tweets
+    if (tweets.length > LIMIT) {
+      searchId = null;
+    }
+  } while (searchId);
+
+  return tweets
+    .reverse()
+    .map(tweet => tweet.text)
+    .join('\n---\n');
+}
+
+/**
+ * Given some text try to narrow down the bill to focus on.
+ *
+ * If you get a bill title.
+ * then we can filter the embeddings to that title
+ * and then try to search for user question for embeddings
+ *
+ * If you do not get a bill title
+ * we are out of luck and we just use the user question to search all the embeddings
+ */
+export async function getBillContext({
+  text,
+  question,
+}: {
+  text: string;
+  question: string;
+}) {
+  const LIMIT = 10;
+  const billTitleResult = await generateText({
+    model: openai('gpt-4o'),
+    temperature: 0,
+    messages: [
+      {
+        role: 'system',
+        content: EXTRACT_BILL_TITLE_PROMPT,
+      },
+
+      {
+        role: 'user',
+        content: text,
+      },
+    ],
+  });
+
+  const questionEmbedding = await generateEmbedding(question);
+  const embeddingArrayString = JSON.stringify(questionEmbedding);
+
+  const billTitle = billTitleResult.text;
+
+  if (billTitle.startsWith('NO_TITLE_FOUND')) {
+    const vectorSearch = await db
+      .select({
+        text: billVector.text,
+        bill: billVector.bill,
+        distance: sql`vector_distance_cos(${billVector.vector}, vector32(${embeddingArrayString}))`,
+      })
+      .from(billVector)
+      .orderBy(
+        // ascending order
+        sql`vector_distance_cos(${billVector.vector}, vector32(${embeddingArrayString})) ASC`,
+      )
+      .limit(LIMIT);
+
+    return vectorSearch.map(row => row.text).join('\n---\n');
+  }
+
+  const billSearch = await db
+    .select({ id: bill.id })
+    .from(bill)
+    .where(like(bill.title, billTitle))
+    .limit(LIMIT);
+
+  const vectorSearch = await db
+    .select({
+      text: billVector.text,
+      bill: billVector.bill,
+      distance: sql`vector_distance_cos(${billVector.vector}, vector32(${embeddingArrayString}))`,
+    })
+    .from(billVector)
+    .where(
+      inArray(
+        billVector.bill,
+        billSearch.map(row => row.id),
+      ),
+    )
+    .orderBy(
+      // ascending order
+      sql`vector_distance_cos(${billVector.vector}, vector32(${embeddingArrayString})) ASC`,
+    )
+    .limit(10);
+
+  return vectorSearch.map(row => row.text).join('\n---\n');
+}
+
 /**
  * Fetches tweets from the Twitter API and queues them for processing.
  */
@@ -138,12 +258,6 @@ export const executeTweets = inngest.createFunction(
           throw new NonRetriableError(e);
         });
 
-        const mainTweet = await getTweet({
-          id: tweetToActionOn.inReplyToId!,
-        }).catch(e => {
-          throw new NonRetriableError(e);
-        });
-
         const question = await step.run('extract-question', async () => {
           const text = tweetToActionOn.text;
           const result = await generateText({
@@ -168,32 +282,20 @@ export const executeTweets = inngest.createFunction(
           return result.text;
         });
 
-        const relevantContext = await step.run(
-          'fetch-relevant-context',
-          async () => {
-            // TODO: rerank the embeddings
-            const questionEmbedding = await generateEmbedding(question);
-            const embeddingArrayString = JSON.stringify(questionEmbedding);
-
-            const vectorSearch = await db
-              .select({
-                id: billVector.id,
-                text: billVector.text,
-                bill: billVector.bill,
-                distance: sql`vector_distance_cos(${billVector.vector}, vector32(${embeddingArrayString}))`,
-              })
-              .from(billVector)
-              .orderBy(
-                // ascending order
-                sql`vector_distance_cos(${billVector.vector}, vector32(${embeddingArrayString})) ASC`,
-              )
-              .limit(5);
-
-            return vectorSearch.map(row => row.text).join('\n---\n');
-          },
-        );
-
         const reply = await step.run('generate-reply', async () => {
+          const threadContext = tweetToActionOn.inReplyToId
+            ? await getTweetContext({
+                id: tweetToActionOn.inReplyToId,
+              }).catch(e => {
+                throw new NonRetriableError(e);
+              })
+            : tweetToActionOn.text;
+
+          const relevantContext = await getBillContext({
+            text: threadContext,
+            question,
+          });
+
           const response = await generateText({
             model: xAi('grok-2-1212'),
             temperature: 0,
@@ -204,11 +306,11 @@ export const executeTweets = inngest.createFunction(
               },
               {
                 role: 'user',
-                content: relevantContext,
+                content: `Context from X: ${threadContext}`,
               },
               {
                 role: 'user',
-                content: mainTweet.text,
+                content: `Context from database: ${relevantContext}`,
               },
               {
                 role: 'user',
@@ -224,7 +326,6 @@ export const executeTweets = inngest.createFunction(
           return response.text;
         });
 
-        // send reply
         const repliedTweet = await step.run('send-tweet', async () => {
           const resp = await twitterClient.v2.tweet(reply, {
             reply: {
@@ -235,12 +336,6 @@ export const executeTweets = inngest.createFunction(
           return {
             id: resp.data.id,
           };
-
-          // await rejectedTweet({
-          //   tweetId: tweetToActionOn.id,
-          //   tweetUrl: tweetToActionOn.url,
-          //   reason: reply,
-          // });
         });
 
         await step.run('notify-discord', async () => {
@@ -264,12 +359,6 @@ export const executeTweets = inngest.createFunction(
             .insert(messageDbSchema)
             .values([
               {
-                text: mainTweet.text,
-                chat: chat.id,
-                role: 'assistant',
-                tweetId: mainTweet.id,
-              },
-              {
                 text: tweetToActionOn.text,
                 chat: chat.id,
                 role: 'user',
@@ -287,26 +376,10 @@ export const executeTweets = inngest.createFunction(
               tweetId: messageDbSchema.tweetId,
             });
 
-          const [chunkMain, chunkActionTweet, chunkReply] = await Promise.all([
-            textSplitter.splitText(mainTweet.text),
+          const [chunkActionTweet, chunkReply] = await Promise.all([
             textSplitter.splitText(tweetToActionOn.text),
             textSplitter.splitText(reply),
           ]);
-
-          const mainEmbeddings = await generateEmbeddings(chunkMain);
-
-          await db.insert(messageVector).values(
-            chunkMain
-              .map((value, index) => ({
-                value,
-                embedding: mainEmbeddings[index],
-              }))
-              .map(({ value, embedding }) => ({
-                message: message[0].id,
-                text: value,
-                vector: sql`vector32(${JSON.stringify(embedding)})`,
-              })),
-          );
 
           const actionTweetEmbeddings =
             await generateEmbeddings(chunkActionTweet);
@@ -318,7 +391,7 @@ export const executeTweets = inngest.createFunction(
                 embedding: actionTweetEmbeddings[index],
               }))
               .map(({ value, embedding }) => ({
-                message: message[1].id,
+                message: message[0].id,
                 text: value,
                 vector: sql`vector32(${JSON.stringify(embedding)})`,
               })),
@@ -333,7 +406,7 @@ export const executeTweets = inngest.createFunction(
                 embedding: replyEmbeddings[index],
               }))
               .map(({ value, embedding }) => ({
-                message: message[2].id,
+                message: message[1].id,
                 text: value,
                 vector: sql`vector32(${JSON.stringify(embedding)})`,
               })),
