+++ agent/src/twitter/prompts.ts
@@ -65,3 +65,5 @@ Response: There is no five-year sunset provision inside the Carla Walker Act (H.
 export const QUESTION_EXTRACTOR_SYSTEM_PROMPT = `You are an advanced text analysis assistant. Your task is to extract the question from a given piece of text. If the text contains a direct question, return it verbatim. If the text is a statement implying a question, rephrase it into a clear and concise question. If no question is present, respond with "NO_QUESTION_DETECTED" Ensure that your output is only the extracted or rephrased question, without additional commentary.`;
 
 export const ENGAGEMENT_DECISION_PROMPT = `Your role is to determine whether a reply to a tweet warrants engagement ("ENGAGE") or should be ignored ("IGNORE"). Your responses must consist of only one word: either "ENGAGE" or "IGNORE." Do not include any analysis, commentary, or reasoning—just output the decision. To decide, analyze replies based on three criteria: connection to context, presence of a clear question, and good-faith indicators. For example, if the reply directly relates to the tweet’s topic and includes a question demonstrating curiosity or seeking clarification, respond with "ENGAGE." Otherwise, respond with "IGNORE." Examples of replies to engage with include: Tweet: "The government is wasting $50 million on forensic DNA gadgets while ignoring community crime prevention programs," Reply: "Why spend $50 million on gadgets instead of reducing crime at the root?" → ENGAGE, Reply: "Can the government explain how this spending benefits public safety?" → ENGAGE; Tweet: "Another $25M for DNA equipment grants. Couldn’t that money fund something more impactful?" Reply: "How do DNA equipment grants actually reduce crime rates?" → ENGAGE, Reply: "Does anyone know how much of this spending goes toward administrative costs?" → ENGAGE. Examples of replies to ignore include: Reply: "Love your project! 🚀 Let’s connect!" → IGNORE, Reply: "Cool post, man! Keep it up!" → IGNORE, Reply: "@elonmusk" → IGNORE, Reply: "This is interesting!" → IGNORE. Analyze tone for good-faith indicators—spam-like or vague replies, such as "Let’s collaborate!" or "Nice work," should always be ignored. For borderline cases, such as "I like parks more than war," use the absence of a clear question to decide on "IGNORE." In all cases, output only "ENGAGE" or "IGNORE," ensuring clarity, brevity, and precision.`;
+
+export const EXTRACT_BILL_TITLE_PROMPT = `You are an AI specialized in analyzing tweets related to U.S. Congressional bills. Given a tweet, extract the official title of the bill mentioned. If multiple bills are referenced, list all their titles. If no bill is mentioned, respond with 'NO_TITLE_FOUND.' Return only the title(s) without additional commentary.`;
